requirements.txt
sentence-transformers
streamlit
google-generativeai
python-dotenv
transformers 
torch 
librosa
streamlit-audiorec
requests
Flask